Refactor theme handling: rename components and implement new ModeToggle and ThemeProvider
